bin/requirements.txt
docutils==0.15.2
lockfile==0.12.2
Pillow==10.1.0
python-daemon==3.0.1
youtube-dl==2021.12.17
